Add `instance` model element
Now, it is possible to have the basis for sequence and object diagrams:
`instance`s.
An instance is similar to a UML class, except it has no properties or
ports but instead an optional instance name (`bob:user` or simply `user`).

Additionally split the logic for the auto-registered classifier name and
the displayed component title.